requirements.txt
beautifulsoup4
Pillow
requests
websocket-client